extract path translation into a concern
for separate testing
