+++ spec/models/concerns/path_translation_spec.rb
@@ -1,0 +1,11 @@
+require "rails_helper"
+
+RSpec.describe PathTranslation do
+  let(:model) { create(:model) }
+
+  context "when creating path from model details" do
+    it "creates a formatted path" do
+      expect(model.formatted_path).not_to be_nil
+    end
+  end
+end
